Move issueReporterUtil to platform, fixes #94505
