+++ src/vs/code/electron-browser/issue/test/testReporterModel.test.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert';
 import { IssueReporterModel } from 'vs/code/electron-browser/issue/issueReporterModel';
-import { normalizeGitHubUrl } from 'vs/code/common/issue/issueReporterUtil';
+import { normalizeGitHubUrl } from 'vs/platform/issue/common/issueReporterUtil';
 import { IssueType } from 'vs/platform/issue/node/issue';
 
 suite('IssueReporter', () => {
