+++ src/vs/platform/issue/common/issueReporterUtil.ts
